added a button to keep the intermediate xmi file
git-svn-id: https://scm.gforge.inria.fr/anonscm/svn/kermeta/trunk@8797 3b3ea266-f0f1-470e-95cb-5b2c9abc2e95

Former-commit-id: 2b88b80af74a41051fee697a6406cc7c4417dc6a

+++ ocl_projects/fr.irisa.triskell.kermeta.ocl/src/fr/irisa/triskell/kermeta/ocl/ui/popup/actions/OCLParseWizard.java
@@ -217,16 +217,17 @@ public class OCLParseWizard extends Wizard {
 			
 			
 			//IFile outfile = inputFile.getWorkspace().getRoot().getFile(inputFile.getLocation().removeFileExtension().addFileExtension("fixed.istecq"));
-			IFile tmpfile = ResourceHelper.getIFile(kmtFile.getFullPath().removeFileExtension().addFileExtension("xmi").toString(),false);
-			if (tmpfile.exists())
-				try {
-					tmpfile.delete(true, null);
-				} catch (CoreException e1) {
-					console.println(new ThrowableMessage(e1));
-					e1.printStackTrace();
-				}
-				
 			
+			if(!outputPage.keepIntermediateXMI()){
+				IFile tmpfile = ResourceHelper.getIFile(kmtFile.getFullPath().removeFileExtension().addFileExtension("xmi").toString(),false);
+				if (tmpfile.exists())
+					try {
+						tmpfile.delete(true, null);
+					} catch (CoreException e1) {
+						console.println(new ThrowableMessage(e1));
+						e1.printStackTrace();
+					}				
+			}
 			
 			// Here I need to redirect Kermeta Console !
